Move command models and rename interaction modules (#701)
* Rename interaction modules
* Move Command to hikari.commands
