+++ hikari/api/interaction_server.py
@@ -30,9 +30,9 @@ import typing
 
 if typing.TYPE_CHECKING:
     from hikari.api import special_endpoints
-    from hikari.interactions import bases as interaction_bases
+    from hikari.interactions import base_interactions
 
-    _InteractionT = typing.TypeVar("_InteractionT", bound=interaction_bases.PartialInteraction, covariant=True)
+    _InteractionT = typing.TypeVar("_InteractionT", bound=base_interactions.PartialInteraction, covariant=True)
     _ResponseT = typing.TypeVar("_ResponseT", bound=special_endpoints.InteractionResponseBuilder, covariant=True)
     _MessageResponseBuilderT = typing.Union[
         special_endpoints.InteractionDeferredBuilder, special_endpoints.InteractionMessageBuilder
@@ -43,7 +43,7 @@ ListenerT = typing.Callable[["_InteractionT"], typing.Awaitable["_ResponseT"]]
 """Type hint of a Interaction server's listener callback.
 
 This should be an async callback which takes in one positional argument which
-subclases `hikari.interactions.bases.PartialInteraction` and may return an
+subclases `hikari.interactions.base_interactions.PartialInteraction` and may return an
 instance of the relevant `hikari.api.special_endpoints.InteractionResponseBuilder`
 subclass for the provided interaction type which will instruct the server on how
 to respond.
@@ -145,12 +145,12 @@ class InteractionServer(abc.ABC):
 
         Parameters
         ----------
-        interaction_type : typing.Type[hikari.interactions.bases.PartialInteraction]
+        interaction_type : typing.Type[hikari.interactions.base_interactions.PartialInteraction]
             Type of the interaction to get the registered listener for.
 
         Returns
         -------
-        typing.Optional[ListenersT[hikari.interactions.bases.PartialInteraction, hikari.api.special_endpoints.InteractionResponseBuilder]
+        typing.Optional[ListenersT[hikari.interactions.base_interactions.PartialInteraction, hikari.api.special_endpoints.InteractionResponseBuilder]
             The callback registered for the provided interaction type if found,
             else `builtins.None`.
         """  # noqa E501 - Line too long
@@ -182,9 +182,9 @@ class InteractionServer(abc.ABC):
 
         Parameters
         ----------
-        interaction_type : typing.Type[hikari.interactions.bases.PartialInteraction]
+        interaction_type : typing.Type[hikari.interactions.base_interactions.PartialInteraction]
             The type of interaction this listener should be registered for.
-        listener : typing.Optional[ListenerT[hikari.interactions.bases.PartialInteraction, hikari.api.special_endpoints.InteractionResponseBuilder]]
+        listener : typing.Optional[ListenerT[hikari.interactions.base_interactions.PartialInteraction, hikari.api.special_endpoints.InteractionResponseBuilder]]
             The asynchronous listener callback to set or `builtins.None` to
             unset the previous listener.
