+++ hikari/interactions/__init__.pyi
@@ -1,5 +1,5 @@
 # DO NOT MANUALLY EDIT THIS FILE!
 # This file was automatically generated by `nox -s generate-stubs`
 
-from hikari.interactions.bases import *
-from hikari.interactions.commands import *
+from hikari.interactions.base_interactions import *
+from hikari.interactions.command_interactions import *
