+++ tests/hikari/interactions/test_command_interactions.py
@@ -26,8 +26,8 @@ from hikari import channels
 from hikari import snowflakes
 from hikari import traits
 from hikari import undefined
-from hikari.interactions import bases
-from hikari.interactions import commands
+from hikari.interactions import base_interactions
+from hikari.interactions import command_interactions
 
 
 @pytest.fixture()
@@ -35,97 +35,13 @@ def mock_app():
     return mock.Mock(traits.CacheAware, rest=mock.AsyncMock())
 
 
-class TestCommand:
-    @pytest.fixture()
-    def mock_command(self, mock_app):
-        return commands.Command(
-            app=mock_app,
-            id=snowflakes.Snowflake(34123123),
-            application_id=snowflakes.Snowflake(65234123),
-            name="Name",
-            description="very descript",
-            options=[],
-            guild_id=snowflakes.Snowflake(31231235),
-        )
-
-    @pytest.mark.asyncio()
-    async def test_fetch_self(self, mock_command, mock_app):
-        result = await mock_command.fetch_self()
-
-        assert result is mock_app.rest.fetch_application_command.return_value
-        mock_app.rest.fetch_application_command.assert_awaited_once_with(65234123, 34123123, 31231235)
-
-    @pytest.mark.asyncio()
-    async def test_fetch_self_when_guild_id_is_none(self, mock_command, mock_app):
-        mock_command.guild_id = None
-
-        result = await mock_command.fetch_self()
-
-        assert result is mock_app.rest.fetch_application_command.return_value
-        mock_app.rest.fetch_application_command.assert_awaited_once_with(65234123, 34123123, undefined.UNDEFINED)
-
-    @pytest.mark.asyncio()
-    async def test_edit_without_optional_args(self, mock_command, mock_app):
-        result = await mock_command.edit()
-
-        assert result is mock_app.rest.edit_application_command.return_value
-        mock_app.rest.edit_application_command.assert_awaited_once_with(
-            65234123,
-            34123123,
-            31231235,
-            name=undefined.UNDEFINED,
-            description=undefined.UNDEFINED,
-            options=undefined.UNDEFINED,
-        )
-
-    @pytest.mark.asyncio()
-    async def test_edit_with_optional_args(self, mock_command, mock_app):
-        mock_option = object()
-        result = await mock_command.edit(name="new name", description="very descrypt", options=[mock_option])
-
-        assert result is mock_app.rest.edit_application_command.return_value
-        mock_app.rest.edit_application_command.assert_awaited_once_with(
-            65234123, 34123123, 31231235, name="new name", description="very descrypt", options=[mock_option]
-        )
-
-    @pytest.mark.asyncio()
-    async def test_edit_when_guild_id_is_none(self, mock_command, mock_app):
-        mock_command.guild_id = None
-
-        result = await mock_command.edit()
-
-        assert result is mock_app.rest.edit_application_command.return_value
-        mock_app.rest.edit_application_command.assert_awaited_once_with(
-            65234123,
-            34123123,
-            undefined.UNDEFINED,
-            name=undefined.UNDEFINED,
-            description=undefined.UNDEFINED,
-            options=undefined.UNDEFINED,
-        )
-
-    @pytest.mark.asyncio()
-    async def test_delete(self, mock_command, mock_app):
-        await mock_command.delete()
-
-        mock_app.rest.delete_application_command.assert_awaited_once_with(65234123, 34123123, 31231235)
-
-    @pytest.mark.asyncio()
-    async def test_delete_when_guild_id_is_none(self, mock_command, mock_app):
-        mock_command.guild_id = None
-
-        await mock_command.delete()
-
-        mock_app.rest.delete_application_command.assert_awaited_once_with(65234123, 34123123, undefined.UNDEFINED)
-
-
 class TestCommandInteraction:
     @pytest.fixture()
     def mock_command_interaction(self, mock_app):
-        return commands.CommandInteraction(
+        return command_interactions.CommandInteraction(
             app=mock_app,
             id=snowflakes.Snowflake(2312312),
-            type=bases.InteractionType.APPLICATION_COMMAND,
+            type=base_interactions.InteractionType.APPLICATION_COMMAND,
             channel_id=snowflakes.Snowflake(3123123),
             guild_id=snowflakes.Snowflake(5412231),
             member=object(),
@@ -144,14 +60,16 @@ class TestCommandInteraction:
         builder = mock_command_interaction.build_response()
 
         assert builder is mock_app.rest.interaction_message_builder.return_value
-        mock_app.rest.interaction_message_builder.assert_called_once_with(bases.ResponseType.MESSAGE_CREATE)
+        mock_app.rest.interaction_message_builder.assert_called_once_with(base_interactions.ResponseType.MESSAGE_CREATE)
 
     def test_build_deferred_response(self, mock_command_interaction, mock_app):
         mock_app.rest.interaction_deferred_builder = mock.Mock()
         builder = mock_command_interaction.build_deferred_response()
 
         assert builder is mock_app.rest.interaction_deferred_builder.return_value
-        mock_app.rest.interaction_deferred_builder.assert_called_once_with(bases.ResponseType.DEFERRED_MESSAGE_CREATE)
+        mock_app.rest.interaction_deferred_builder.assert_called_once_with(
+            base_interactions.ResponseType.DEFERRED_MESSAGE_CREATE
+        )
 
     @pytest.mark.asyncio()
     async def test_fetch_initial_response(self, mock_command_interaction, mock_app):
@@ -165,7 +83,7 @@ class TestCommandInteraction:
         mock_embed_1 = object()
         mock_embed_2 = object()
         await mock_command_interaction.create_initial_response(
-            bases.ResponseType.MESSAGE_CREATE,
+            base_interactions.ResponseType.MESSAGE_CREATE,
             "content",
             tts=True,
             embed=mock_embed_1,
@@ -179,7 +97,7 @@ class TestCommandInteraction:
         mock_app.rest.create_interaction_response.assert_awaited_once_with(
             2312312,
             "httptptptptptptptp",
-            bases.ResponseType.MESSAGE_CREATE,
+            base_interactions.ResponseType.MESSAGE_CREATE,
             "content",
             tts=True,
             flags=64,
@@ -192,12 +110,12 @@ class TestCommandInteraction:
 
     @pytest.mark.asyncio()
     async def test_create_initial_response_without_optional_args(self, mock_command_interaction, mock_app):
-        await mock_command_interaction.create_initial_response(bases.ResponseType.DEFERRED_MESSAGE_CREATE)
+        await mock_command_interaction.create_initial_response(base_interactions.ResponseType.DEFERRED_MESSAGE_CREATE)
 
         mock_app.rest.create_interaction_response.assert_awaited_once_with(
             2312312,
             "httptptptptptptptp",
-            bases.ResponseType.DEFERRED_MESSAGE_CREATE,
+            base_interactions.ResponseType.DEFERRED_MESSAGE_CREATE,
             undefined.UNDEFINED,
             flags=undefined.UNDEFINED,
             tts=undefined.UNDEFINED,
